refactor(api): flatten src/lib into src/
The lib/ directory held only firebase/ and, since #1149, catalogue.ts.
Neither grouping earned the extra nesting: every other src/ subdirectory
sits at the top level, so the wrapper only made imports longer and the
Codecov component name misleading ("API: Lib" against firebase content).

Move firebase/ to src/firebase/ and catalogue.ts to src/catalogue.ts,
rewrite the @/lib/... imports, and repoint the api-firebase Codecov
component at the new path.

Closes #463
