+++ apps/api/src/firebase/app.ts
