+++ apps/api/src/repositories/profile/index.ts
@@ -3,7 +3,7 @@
 
 import type { ISOTimestamp, ProfileUpdate, UserProfile } from '@genshin/domain';
 
-import { db } from '@/lib/firebase/firestore.js';
+import { db } from '@/firebase/firestore.js';
 
 import { fromDocument, toDocument } from './document.js';
