+++ apps/api/src/repositories/teams/index.ts
@@ -3,7 +3,7 @@
 
 import type { CollectionTeam, ISOTimestamp, TeamSlot } from '@genshin/domain';
 
-import { db } from '@/lib/firebase/firestore.js';
+import { db } from '@/firebase/firestore.js';
 
 import { fromDocument, toDocument } from './document.js';
