+++ apps/api/src/routes/weapons.test.ts
@@ -8,13 +8,13 @@ import { getWeaponById } from '@genshin/game-data';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { app } from '@/app.js';
-import { verifyToken } from '@/lib/firebase/auth.js';
+import { verifyToken } from '@/firebase/auth.js';
 import { toMediaTypeString } from '@/middleware/negotiate-content.js';
 import { weaponItemV1 } from '@/profiles/alps/weapon/item-v1.js';
 import * as Weapons from '@/repositories/weapons/index.js';
 import { FAKE_TOKEN, authedRequest } from '@/test/auth-requests.js';
 
-vi.mock('@/lib/firebase/auth.js', () => ({
+vi.mock('@/firebase/auth.js', () => ({
   verifyToken: vi.fn(),
 }));
