+++ codecov.yml
@@ -85,7 +85,7 @@ component_management:
     - component_id: api-firebase
       name: 'API: Firebase'
       paths:
-        - apps/api/src/lib/firebase/**
+        - apps/api/src/firebase/**
       flag_regexes:
         - api
     - component_id: game-data-versions
